+++ src/types.ts
@@ -1,0 +1,8 @@
+export interface TokenData {
+  chainId: number;
+  address: string;
+  name: string;
+  symbol: string;
+  decimals: number;
+  logoURI: string;
+}
